# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Backend for AtlasCV
host: localhost:10013
basePath: /
schemes:
  # tip: remove http to make production-grade
  - http
  - https
consumes:
  - application/json
  - text/html
  - application/x-www-form-urlencoded
produces:
  - application/json
  - text/html
# securityDefinitions:
#   applicantsAndEmployers:
#     name: Authorization
#     type: apiKey
#     in: header
#     description: applicants and employers
#   applicantsOnly:
#     name: Authorization
#     type: apiKey
#     in: header
#     description: applicant accounts only
#   employersOnly:
#     name: Authorization
#     type: apiKey
#     in: header
#     description: employer accounts only
#   anyUser:
#     name: Authorization
#     type: apiKey
#     in: header
#     description: any user
#   admin:
#     name: Authorization
#     type: apiKey
#     in: header
#     description: AtlasCV admin account
#   adminAndEmployers:
#     name: Authorization
#     type: apiKey
#     in: header
#     description: AtlasCV admins and employers
paths:
  /applicants:
    x-swagger-router-controller: applicants
    get:
      description: Get all applicants
      operationId: getAllApplicants
      responses:
        "200":
          description: Successfully return all applicants
          schema:
            type: object
        "500":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      x-swagger-router-controller: applicants
      description: Creates a new applicant
      operationId: createApplicant
      parameters:
        - name: data
          in: body
          description: data required to create an applicant
          required: true
          schema:
            type: object
            required:
              - firstName
              - lastName
              - email
              - password
            properties:
              firstName:
                type: string
              lastName:
                type: string
              email:
                type: string
              password:
                type: string
              uuid:
                type: string
      responses:
        "201":
          description: Successfully create a new applicant
          schema:
            type: object
        "500":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /applicants/{applicantId}:
    x-swagger-router-controller: applicants
    get:
      description: Get all applicants
      operationId: getApplicantById
      parameters:
        - name: applicantId
          in: path
          description: id of the applicant being retrieved
          required: true
          type: string
      responses:
        "200":
          description: Successfully return applicant by id
          schema:
            type: object
        "500":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      description: update an existing applicant
      operationId: updateApplicant
      parameters:
        - name: applicantId
          in: path
          description: id of the applicant being updated
          required: true
          type: string
        - name: data
          in: body
          description: data required to create an applicant
          required: true
          schema:
            type: object
            properties:
              firstName:
                type: string
              lastName:
                type: string
              email:
                type: string
              city:
                type: string
              profileImgUrl:
                type: string
              transcript:
                type: string
              aboutMe:
                type: string
              gender:
                type: string
              phone:
                type: string
              birthday:
                type: string
              password:
                type: string
      responses:
        "200":
          description: Successfully update an existing applicant
          schema:
            type: object
        "500":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /applicants/{applicantId}/skills:
    post:
      x-swagger-router-controller: skills
      description: Adds a skill for an applicant
      operationId: addSkillsToApplicant
      parameters:
        - name: applicantId
          in: path
          description: id of the applicant's skills being added
          required: true
          type: string
        - name: data
          in: body
          description: data required to create an applicant
          required: true
          schema:
            type: object
            required:
              - skill
            properties:
              skill:
                type: object
                properties:
                  id:
                    type: number
                  yearsExperience:
                    type: string
      responses:
        "201":
          description: Successfully create a new applicant
          schema:
            type: object
        "500":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      x-swagger-router-controller: skills
      description: removes a skill from an applicant
      operationId: removeSkillFromApplicant
      parameters:
        - name: applicantId
          in: path
          description: id of the applicant's skills being removed
          required: true
          type: string
        - name: data
          in: body
          description: skillId to remove
          required: true
          schema:
            type: object
            required:
              - skillId
            properties:
              skillId:
                type: number
      responses:
        "201":
          description: Successfully create a new applicant
          schema:
            type: object
        "500":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /applicants/{applicantId}/certifications:
    post:
      x-swagger-router-controller: certifications
      description: Adds a certification for an applicant
      operationId: addSkillsToApplicant
      parameters:
        - name: applicantId
          in: path
          description: id of the applicant's certifications being added
          required: true
          type: string
        - name: data
          in: body
          description: data required to create an applicant
          required: true
          schema:
            type: object
            required:
              - certification
            properties:
              certification:
                type: object
                properties:
                  id:
                    type: number
                  yearsExperience:
                    type: string
      responses:
        "201":
          description: Successfully create a new applicant
          schema:
            type: object
        "500":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      x-swagger-router-controller: certifications
      description: removes a certification from an applicant
      operationId: removeSkillFromApplicant
      parameters:
        - name: applicantId
          in: path
          description: id of the applicant's certifications being removed
          required: true
          type: string
        - name: data
          in: body
          description: certificationId to remove
          required: true
          schema:
            type: object
            required:
              - certificationId
            properties:
              certificationId:
                type: number
      responses:
        "201":
          description: Successfully create a new applicant
          schema:
            type: object
        "500":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /applicants/{applicantId}/industrySectors:
    post:
      x-swagger-router-controller: industrySectors
      description: Adds an industry sector for an applicant
      operationId: addIndustrySectorsToApplicant
      parameters:
        - name: applicantId
          in: path
          description: id of the applicant's industrySectors being added
          required: true
          type: string
        - name: data
          in: body
          description: data required to create an applicant
          required: true
          schema:
            type: object
            required:
              - industrySector
            properties:
              industrySector:
                type: object
                properties:
                  id:
                    type: number
                  yearsExperience:
                    type: string
      responses:
        "201":
          description: Successfully create a new applicant
          schema:
            type: object
        "500":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      x-swagger-router-controller: industrySectors
      description: removes a industrySector from an applicant
      operationId: removeIndustrySectorFromApplicant
      parameters:
        - name: applicantId
          in: path
          description: id of the applicant's industrySectors being removed
          required: true
          type: string
        - name: data
          in: body
          description: industrySectorId to remove
          required: true
          schema:
            type: object
            required:
              - industrySectorId
            properties:
              industrySectorId:
                type: number
      responses:
        "201":
          description: Successfully create a new applicant
          schema:
            type: object
        "500":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /applicants/{applicantId}/industries:
    post:
      x-swagger-router-controller: industries
      description: Adds a desired industries for an applicant
      operationId: addIndustriesToApplicant
      parameters:
        - name: applicantId
          in: path
          description: id of the applicant's skills being added
          required: true
          type: string
        - name: data
          in: body
          description: data required to create an applicant
          required: true
          schema:
            type: object
            required:
              - industryIds
            properties:
              industryIds:
                type: array
                items:
                  type: number
      responses:
        "201":
          description: Successfully create a new applicant
          schema:
            type: object
        "500":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /applicants/{applicantId}/education:
    post:
      x-swagger-router-controller: education
      description: Adds education experience
      operationId: addEducationToApplicant
      parameters:
        - name: applicantId
          in: path
          description: id of the applicant's skills being added
          required: true
          type: string
        - name: data
          in: body
          description: data required to create education for an applicant
          required: true
          schema:
            type: object
            required:
              - areaOfStudy
              - gpa
              - educationLevel
              - university
            properties:
              areaOfStudy:
                type: string
              gpa:
                type: string
              educationLevel:
                type: string
              university:
                type: string
      responses:
        "201":
          description: Successfully create and associate education with applicant
          schema:
            type: object
        "500":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /applicants/{applicantId}/jobExperience:
    post:
      x-swagger-router-controller: jobExperience
      description: Adds job experience
      operationId: addJobExperienceToApplicant
      parameters:
        - name: applicantId
          in: path
          description: id of the applicant's skills being added
          required: true
          type: string
        - name: data
          in: body
          description: data required to create job experience for an applicant
          required: true
          schema:
            type: object
            required:
              - name
              - numOfYears
              - currentlyWorkingHere
              - companyName
            properties:
              name:
                type: string
              numOfYears:
                type: number
              currentlyWorkingHere:
                type: boolean
              companyName:
                type: string
              description:
                type: string
      responses:
        "201":
          description: Successfully create and associate job experience with applicant
          schema:
            type: object
        "500":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /personality-evaluations:
    post:
      x-swagger-router-controller: personalityEvaluations
      description: Creates an instance of a personality evaluations on the backend for the potential applicant to be associated with an applicant later.
      operationId: createPersonalityEvaluation
      parameters:
        - name: data
          in: body
          description: empty object
          required: true
          schema:
            type: object
      responses:
        "201":
          description: Successfully create results
          schema:
            type: object
        "500":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /personality-evaluations/{uuid}:
    get:
      x-swagger-router-controller: personalityEvaluations
      description: Gets an instance of a personality evaluation using the uuid
      operationId: getPersonalityEvaluation
      parameters:
        - name: uuid
          in: path
          required: true
          description: the
          type: string
      responses:
        "200":
          description: Successfully return the personality evaluation
          schema:
            type: object
        "500":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      x-swagger-router-controller: personalityEvaluations
      description: Gets an instance of a personality evaluation using the uuid
      operationId: updatePersonalityEvaluation
      parameters:
        - name: uuid
          in: path
          required: true
          description: the
          type: string
        - name: data
          in: body
          description: data to update a personality evaluation
          required: true
          schema:
            type: object
            properties:
              answers:
                type: array
                items:
                  type: number
              currentQuestionIndex:
                type: number
              scoreSignature:
                type: string
              completed:
                type: boolean
      responses:
        "200":
          description: Successfully return the updated personality evaluation
          schema:
            type: object
        "500":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /user/{userId}/confirm:
    put:
      x-swagger-router-controller: users
      description: Confirms the signup for a user with their password
      operationId: confirmUser
      parameters:
        - name: userId
          in: path
          required: true
          description: the
          type: number
        - name: data
          in: body
          description: data required to confirm a user
          required: true
          schema:
            type: object
            required:
              - password
            properties:
              password:
                type: string
      responses:
        "200":
          description: Successfully confirm a user with a password
          schema:
            type: object
        "500":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /login:
    post:
      x-swagger-router-controller: users
      description: logs in the user
      operationId: login
      parameters:
        - name: data
          in: body
          description: data required to log in a user
          required: true
          schema:
            type: object
            required:
              - email
              - password
            properties:
              email:
                type: string
              password:
                type: string
      responses:
        "200":
          description: Successfully confirm a user with a password
          schema:
            type: object
        "500":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /industries:
    get:
      x-swagger-router-controller: industries
      description: Gets all of the industries in the database
      operationId: getAllIndustries
      responses:
        "200":
          description: Successfully return all industries
          schema:
            type: object
        "500":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /industrySectors:
    get:
      x-swagger-router-controller: industrySectors
      description: Gets all of the industrySectors in the database
      operationId: getAllIndustrySectors
      responses:
        "200":
          description: Successfully return all industrySectors
          schema:
            type: object
        "500":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /industrySectors/{industryId}:
    get:
      x-swagger-router-controller: industrySectors
      description: Get an individual industry sector's information
      operationId: getIndustrySectorsByIndustry
      parameters:
        - name: industryId
          in: path
          required: true
          description: the
          type: string
      responses:
        "200":
          description: Successfully return all industrySectors
          schema:
            type: object
        "500":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /skills:
    get:
      x-swagger-router-controller: skills
      description: Gets all of the skills in the database
      operationId: getAllSkills
      responses:
        "200":
          description: Successfully return all skills
          schema:
            type: object
        "500":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /certifications:
    get:
      x-swagger-router-controller: certifications
      description: Gets all of the certifications in the database
      operationId: getAllCertifications
      responses:
        "200":
          description: Successfully return all certifications
          schema:
            type: object
        "500":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /me:
    get:
      x-swagger-router-controller: users
      description: Gets the user's information
      operationId: getMe
      responses:
        "200":
          description: Successfully return your information
          schema:
            type: object
        "500":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
definitions:
  ErrorResponse:
    required:
      - success
      - status
      - message
    properties:
      success:
        type: boolean
      status:
        type: integer
      message:
        type: string